adding xcodetool to Tools
Summary:
xcodetool is this front-end for xcodebuild that I've been working on -
it provides some extra knobs and levers for building/running tests on
the BuildBot side.

It's not totally done yet, but let's stick it in Tools and let people
use it.  It has one helpful feature that lets you run an individual
test case - people seem to be interested in that. (see README)

About the dot files (.xcodetool-args) - these make it easier to invoke
xcodetool without having to type a bunch of params.  It pulls them from
the JSON file as if they were provided on the command-line.

About the binaries under Vendor - I didn't bother to put them under
fbobjc's main Vendor dir since xcodetool will end up on GitHub soon
enough.

More docs & code cleanup needed, but it's useful now.

Test Plan: It can build and run tests for Facebook.