fix hang that could occur while waiting for an NSTask to end.
Summary:
Sometimes xcodetool would hang while trying to read the last bit of
output from an NSTask.

It gets the output from the task via a pipe, and reads until EOF.  The
write-handle of the pipe was sometimes not getting closed, so it would
never get the EOF and the read would just block forever.

Passing the NSPipe to -[NSTask setStandardOutput:] fixes it - NSTask
will automatically close the write-side for us.

Test Plan: ../../Tools/xcodetool/xcodetool.sh build-tests worked.