report tests as failed when they crash.
Summary:
When an OCUnit test crashes, it doesn't tell you it failed - the process
just dies.  We can tell when this happens because we'll see a
'begin-test' event but no 'end-test' event to close it out.

In that case, let's fire off the 'end-test' and 'end-test-suite' events
as if the test process actually did finish.  This way the Reporter(s) have
a sane event stream to display.

Test Plan:
From Tools/xcodetool, these worked --

```
./xcodetool.sh -project xcodetool/xcodetool-tests/TestData/TestProject-Library/TestProject-Library.xcodeproj -scheme TestProject-Library -configuration Debug -arch i386 -sdk iphonesimulator6.1 run-tests
```

```
./xcodetool.sh -project xcodetool/xcodetool-tests/TestData/TestsThatCrash/TestsThatCrash.xcodeproj -scheme TestsThatCrash -configuration Debug -arch i386 -sdk iphonesimulator6.1 run-tests
```