build otest-lib and app-uninstaller for oldest sim SDKs.
Summary:
For otest-lib-ios and app-uninstaller.app, we want to set the
IPHONEOS_DEPLOYMENT_TARGET to be the oldest installed simulator SDK.
This way they're the most compatible - we can use them even if we're
running the 5.0 simulator.

Test Plan: The scripts still work.  Verified they're now targetting the right SDKs.