phabricator reporter should include exception text, if any.
Summary:
If a test fails with an exception, we should include that in 'userdata'
field for the test just like we do with any stdout/stderr.

This way you'll see the exception text for tests when they appear in the
"Unit" section of Phabricator.

Test Plan: Added test.  Manual test looked good.