if failureReason is nil, send @"" instead of [NSNull null]
Summary:
This fixes a crash in xcodetool --
https://phabricator.fb.com/P2092156

There were places in the TextReporter and PhabricatorReporter code that
assumed event[@"failureReason"] was always a string.  Sometimes it was
[NSNull null] instead.

Test Plan:
hard-coded event[@"failureReason"] to be @"" in xcocebuild-lib and
verified that the reporters were cool with an empty string being there -
they were.