pretend test bundle succeeds if no tests ran but exit code was != 0
Summary:
This is a fix for the bug where xcodetool would claim that tests failed
even though all tests passed.

```
** RUN-TESTS FAILED: 1517 of 1517 tests passed ** (216447 ms)
```

This is better explained in the comment but...

This was uncovered because a test target had no source files and so
no test binary was created.

When otest would try to run it, it would exit with non-zero status.

In Xcode, they just ignore the non-zero exit status and claim it
succeeded.

We should to do the same.