Fix build -only to include deps
Summary:
Previously, when calling xctool build-tests with the -only
flag to specify just one test, it would fail to build the dependencies
of the test target. This usually caused the test to fail to build.

Now, we build the test and other non-test targets in the same scheme.

Test Plan:
Unit tests updated. Tested with a few Facebook projects
containing multiple tests.