Override test skipped flag in scheme when specifying -only
Summary:
We have a few workspaces that contain skipped tests.
If the user specifies -only to xctool, we should ignore the
skipped flag.

Test Plan:
Tested on Facebook project with skipped tests.
Previously, -only flag would not build or test them. Now, it does.

Confirmed without -only flag that skipped tests are not built or
run.

@fpotter, mind helping me write a unit test for this?