make tests pass when built from the command line.
Summary:
Some code would follow different paths if we were running under test.
To do that, we'd look at the process name and if it was 'otest', we'd
assume we were under test.

The process name for otest when running Cmd-U in xcode is always 'otest'
- that's nice.

But, when running from the command line, 'otest' can instead be
'otest-x86_64'.  So, let's watch for both.

Test Plan:
This passes --

```
Tools/xctool/build.sh TEST_AFTER_BUILD=YES
```