Let tests use iphonesimulator SDK and i386 arch if not otherwise specified
Summary:
Right now, RunTestsAction doesn't work with the iphoneos SDK.

This is fine, but it requires explicitly specifying an SDK and arch
when running tests. (This is because we use the defaults from the Xcode
project, which are usually iphoneos / armv7 for iOS targets).

In this change, I add -[Action validateSDK:] to let RunTestsAction
specify whether it supports an SDK, and I use it to let us fall
back to iphonesimulator / i386 when running tests.

Test Plan:
Unit tests included. Ran:
  % Tools/xctool/build.sh TEST_AFTER_BUILD=YES
  % Tools/xctool/xctool.sh -find-target FBFoundationTests test