Add -find-target-exclude-paths to skip searching paths
Summary:
There are a jillion workspaces in Users that just don't
build (references to users' home directories on their laptops,
etc.).

We need to exclude them from -find-target, so we don't barf
when trying to find a reference to a project that doesn't exist.

This adds a -find-target-exclude-paths path1:path2:... option
that lets us exclude any paths that include a component that
matches one of the entries in the exclude list.

In a followup diff, I need this so we can properly look at project
references from a workspace.

Test Plan:
 % Tools/xctool/build.sh TEST_AFTER_BUILD=YES
 % Tools/xctool/xctool.sh -find-target FBFoundationTests -find-target-exclude-paths Users:Vendor test