Prefer schemes defined in workspaces to ones defined in projects
Summary:
Many of our schemes do not build as defined in the Xcode projects;
they have to be built as part of a workspace.

If we find a target defined in both a project scheme and a workspace
that contains references to those projects, we'll now prefer
the workspace.

Test Plan:
Unit test updated.
 % Tools/xctool/build.sh TEST_AFTER_BUILD=YES
 % Tools/xctool/xctool.sh -find-target FBFoundationTests -find-target-exclude-paths Users:Vendor clean
 % Tools/xctool/xctool.sh -find-target FBFoundationTests -find-target-exclude-paths Users:Vendor test