Add message reporting with timestamp and level
Summary:
For -find-target I want to log an informational level when I
find the workspace and scheme to build.

In this diff, I add message reporting to the Reporter class, and
tweak the Options class to validate and set up the Reporter early,
so later validation can use the reporter.

I added one reporter message to log the workspace/project and scheme
for -find-target.