Register reporters so we can report messages anywhere
Summary:
I'd like to be able to report messages from code
that may not have access to the reporters array.

This adds a global array of reporters which ReportMessage() now uses
instead of requiring the caller to pass in an array of reporters.

Test Plan: Tools/xctool/build.sh TEST_AFTER_BUILD=YES