New build-tests option -skip-deps
Summary:
In D782123, I changed -only to build dependencies,
which caused builds to get slower for the use-case of editing
a single test / building / running that test.

In this diff, I introduce a new build-tests option -skip-deps
that skips building everything except the test target itself.

When I added tests, I noticed a copy-pasto in BuildTestsActionTests
and fixed RunTestsAction -> BuildTestsAction. (The test was working
because Objective-C message passing is generic and works even
if the type is wrong.)

Test Plan:
Tested from command line and from unit tests:
  % ./Tools/xctool/build.sh TEST_AFTER_BUILD=YES
  % ./Tools/xctool/xctool.sh -workspace Libraries/FBFoundation/FBFoundation.xcworkspace -scheme FBFoundation -sdk iphonesimulator6.1 -arch i386 build-tests -only JSONKitTests -skip-deps