xctool (easy): Use @try / @finally so we always clean up reporters on early exit
Summary:
Missed this in D788268. We need to always clean up in XCTool, even if
we exit early.

Test Plan:
Made a typo in a unit test that caused XCTool option validation
to exit early, which caused every subsequent test to fail when registering
reporters. Now, only the one test fails.