fix issue where iOS application tests sometimes don't start.
We use iPhoneSimulatorRemoteClient.framework to kick the simulator and
start iOS application tests.  We ask iPhoneSimulatorRemoteClient to
start the TEST_HOST app for us and then we insert the test bundle via
DYLD_INSERT_LIBRARIES.  For some reason iPhoneSimulatorRemoteClient does
not reliably set the app launch environment if the app is being freshly
installed.

This workaround makes it so that we always install the TEST_HOST app
once before we try and run it.  By doing this way, the app launch
environment seems to get set right every time.