fix issue where run-tests -only might run only the disabled tests.
If you had disabled some test classes in a scheme and then used -only to
run that test target, it was running only the disabled tests.

Tested by...

Adding a new DisabledTests test class to TestProject-Library and
unchecked it in the scheme.  Then ran...

```
./xctool.sh -project xctool/xctool-tests/TestData/TestProject-Library/TestProject-Library.xcodeproj -scheme TestProject-Library -arch i386 -configuration Debug -sdk iphonesimulator run-tests -only TestProject-LibraryTests
```

... and verified it ran the non-disabled tests.

Also checked that you could still run the disabled test if you really
wanted with...

```
./xctool.sh -project xctool/xctool-tests/TestData/TestProject-Library/TestProject-Library.xcodeproj -scheme TestProject-Library -arch i386 -configuration Debug -sdk iphonesimulator run-tests -only TestProject-LibraryTests:DisabledTests
```