adding Common dir to hold shared code, moving Swizze functions there.
I need to use the same swizzle functions across multiple targets, so
lets stick them in this Common dir.

We're going to just include the source files themselves rather than
making a library out of them - our targets will be a mix of MRR, ARC,
and GC-code, and iOS / Mac, so linking a common library is not possible.