use EventQueue() instead of __eventQueue
We used to initialize __eventQueue in the 'constructor' for our lib, but
instead lets do it on-demand.

It turns out that it's possible one of the write functions we're
interposing gets called before our constructor gets run.  In that
case, the interposed write() was trying to use an uninitialized
EXC_BAD_ACCESS and then crashing.