ReportMessage becomes ReportStatusMessage w/ optional begin & end
Sometimes we want to announce the beginning and ending of a long running
task - e.g. collecting build settings via -showBuildSettings, installing
an app in the simulator for app tests, trying to search for a target
via -find-target.

With ReportMessage as it was, to announce the begin and end of
something, we'd have to write multiple lines. e.g. --

```
[Info] Doing some thing...
[Info] Finsihed doing some thing. (100 ms)
```

Instead, we'd prefer to have the "finish" message write over the "begin"
message so that we first see...

```
[Info] Doing some thing...
```

... but when it finishes, the line gets rewritten to be ...

```
[Info] Finsihed doing some thing. (100 ms)
```

This changes ReportMessage() to ReportStatusMessage(), which like the
original, writes 1 line of output via the text reporter.

But, for things that have a begin and end, use
ReportStatusMessageBegin() and ReportStatusMessageEnd().  The former
will write 1 line of output, the latter will rewrite that 1 line of
output.