fix bug where clean action wasn't cleaning targets set as build for test.
In xctool, we have testables (targets added ot the 'Test' pane of the
scheme editor), and we have buildables for tests (targets added to the
'Build' pane in the scheme editor but with 'Test' checked).

For the clean action, we weren't cleaning the latter (targets marked as
build for test).

Also, might as well unify that logic in XcodeSubjectInfo since we're now
using in the build-tests and clean actions.