errors for run-tests were sometimes being swallowed.
e.g., if the test bundle didn't generate correctly, the output looked
like...

```
  run-test TestProject-LibraryTests.octest (iphonesimulator6.1, logic-test, GC OFF)

** TEST FAILED: 0 of 0 tests passed ** (1167 ms)
```

We had a check that said if the tests failed to run && we got no test
events, pretend things succeeded.  This was to immitate some Xcode
behavior.  But, really, we should only do that when we don't have an
error message.

Now output looks like...

```
  run-test TestProject-LibraryTests.octest (iphonesimulator6.1, logic-test, GC OFF)
  failed: Test bundle not found at:
/Users/fpotter/Library/Developer/Xcode/DerivedData/TestProject-Library-gqcuxcsyguaqwugnnwmftlazxbyg/Build/Products/Release-iphonesimulator/TestProject-LibraryTests.octest
** TEST FAILED: 0 of 0 tests passed ** (1167 ms)
```