can build workspaces w/ projects targeting different SDKs
xctool used to incorrectly assume every target in your workspace was
targeting the same SDK.  At startup, it would find the SDK of the first
target in your scheme, then assume every other target used the same SDK.
This meant that if you had OSX targets and iOS targets in your
workspace, xctool would fail to build because of SDK confusion.  (issue #12)

As part of this, we also have to remove the logic that would
automatically select the 'iphonesimulator' SDK for 'iphoneos' targets.
This has the downside that callers will now have to remember to pass
'-sdk iphonesimulator' when testing iOS apps, but the upside that
callers can actually build device/ARM versions of the test bundles.
This was inevitable - we will eventually have on-device testing.