TemporaryDirectoryForAction() gives you a temp dir for the current action
SchemeGenerator used to create its own temp directory to store the
generated workspace and scheme.  I'm about to add something else that
needs a temp directory and it'll be cleaner if we just have a general
purpose temp dir for the duration of each action.  The directory will
get cleaned up as soon as the action finishes.