fix the issue where we accumulate Tests-* folders in DerivedData
Every time xctool builds or runs tests, it generates a new workspace and
scheme.  When we invoke xcodebuild on that workspace/scheme, xcodebuild
creates a folder in the user's DerivedData like
'Tests-dgtnwkoyuhjfcibwyjiprineykfj'.

Since we override OBJROOT/SYMROOT/etc at the same time, this folder
serves no real purpose - it's junk.

Let's override 'IDECustomDerivedDataLocation' so this junk directory
gets created in the action's temp directory and is automatically
cleaned up on each run.

Test Plan:
Commented out the code that cleans up the temp dir - verified that the
junk DerivedData folder was placed there.