make Options return the XcodeSubjectInfo instance.
The way XcodeSubjectInfo was created was pretty funky.  -[XCTool run]
would alloc/init a XcodeSubjectInfo instance, but it would be unusable
at this point.  Then it would pass that instance to -[Options
validate...], which would fill in some important values in
XcodeSubjectInfo, at which point it would be usable.

This is horribly confusing, and it's actually still pretty bad, but this
is a little better.  Let's have -[Options validate...] return the
instance in a usable state.

Test Plan:
Tests still pass, and I can build/test manually.