GetAvailableSDKsAndAliases() should empty the env before running xcodebuild.
Previoiusly, we never called `setEnvironment:` on the NSTask, and so the
caller's environment would pass through.

Usually that's OK, but when running tests from Xcode,
OBJC_DISABLE_GC=YES will get set and then `xcodebuild` will fail to run
(because it uses GC).

Note that DEVELOPER_DIR can still be used to be override which
`xcodebuild` to use, since `XcodeDeveloperDirPath()` will call out to
`xcodebuild`.