break `-bucketSize` into `-logicTestBucketSize` and `-appTestBucketSize`
So, the odds that you actually want to bucketize and parallelize
application tests are pretty low right now.

There's a lot of overhead in interacting with the simulator, so you want
to minimize the number of times you install/launch apps.  If you
bucketize your app tests, your test runs will take a lot longer.

Later, when we have parallelized simulators, it might make more sense to
bucketize app tests.