fix bug where we'd claim 'Failed to run tests' if any app tests failed.
We used to return the error message 'Failed to run tests' if any test in
the test bundle failed to pass.

That's not right - this error message was meant to communicate failures
in the test infrastructure, not regular test case failures.

We'll now distinguish between test and infra failures when running tests
in the simulator, and we'll only set `*error` for infra-type failures.