fix problem where xctool can only run sim tests for the version of Xcode it was compiled against.
xctool depends on iPhoneSimulatorRemoteClient.framework, which is a
private framework for interacting with the simulator that comes bundled
with Xcode.

Since xctool can work with multiple verstions of Xcode and since each of
these Xcode versions might live at different paths, we don't want to
strongly link iPhoneSimulatorRemoteClient.framework.  e.g., if we linked
to
`/Applications/Xcode.app/.../.../iPhoneSimulatorRemoteClient.framework`
but Xcode was installed elsewhere, xctool would fail to run.

To workaround this, we weak link the framework and at startup, we tweak
the DYLD_FALLBACK_FRAMEWORK_PATH so that it points to the correct paths
for whatever the current version of Xcode is.