Show currently parallel running bundle
In case tests hang, we want to figure out which test is yet running.
This can hard when tests are run in parallel.

This patch updates the info when a bundle finishes.
Note that application tests will not be reported since they never
run in parallel currently. Anyway bundle name is added for the sake
of consistency.