Collapse test suites into a single suite
Multiple levels have test suites is of questionable value and creates a
headache in xctool. For example, when running tests in parallel the test
suite is emitted for each grouping of tests, even if they're within the
same suite.

This change will only emit the first test suite seen, ignoring any
others.