Add support for emitting results for all tests, even after a crash
Previously if a test suite/test case crashed we wouldn't get any output
for remaining tests. This change will instead emit test results for ALL
tests, even those that didn't get a chance to run. If a test didn't run
then we consider it a failure and label it with a message stating so.

To implement this, OCUnitCrashFilter was changed to implement the
reporter class so it can listen to all events that the reporters are
getting. With the help of the new OC*State classes it will keep track of
what the state of the test run is. If the test suite crashes for any
reason then OCUnitCrashFilter will take over emitting results that the
otest-shim would have emitted.