fix the 'Cannot get systemRoot' exception we get when testing on Xcode5
We used to use DTiPhoneSimulatorSystemRoot to fetch the absolute path
to the SDK root.  That's fine, except it throws an exception when a
specific SDK version can't be found.

Instead, let's form the path ourselves - it never seems to change
anyway.

When running under test, it's OK that we pretend to use SDKs that don't
really exist on the system so long as the test uses FakeTaskManager to
intercept/fake the call to otest-query.