force i386 when running `otest` for iOS logic tests.
As of the Xcode 5 GM, `otest` for the iPhoneSimulator SDK became
a universal binary.

The problem with that is that 1) Xcode doesn't yet build iPhoneSimulator
binaries as x86_64 - they're all still i386; and 2) otest-shim + friends
are still i386 only.

For now, we can just force i386 and things will keep working.  At
somepoint Apple is going to start building simulator binaries as x86_64
and we'll have to make otest-shim + otest-query universal.

We'll probably also have to do what Xcode does when testing universal
binaries - i.e. runs tests twice, once as i386 and once as x86_64.