if xctool is interrupted, its child processes should die.
Let's create all tasks using this `CreateTaskInSameProcessGroup()`
function.  This will give us an NSTask that's configured to NOT start a
new process group for every spawned task.  (The default is to spawn all
children in new process groups.)

This way, if we spawn a long running build (for instance), it won't keep
running if someone does ctrl-c on xctool.