make_release.sh should run tests via xctool itself.
We used to use TEST_AFTER_BUILD=YES, but that doesn't work with Xcode 5
anymore.

There is `xcodebuild ... test` now, but that doesn't respect the
`-configuration` parameter.  If we try to run tests this way, the
targets will get built as Release (we pass -configuration Release), but
the tests will get run as if the configuration is Debug (since our
scheme specificies Debug for the configuration).  Seems like a bug in
xcodebuild.

Let's just use xctool to test itself (in the Release configuration).