adding a `-showTasks` option which logs info on all spawned tasks.
When `-showTasks` is passed, xctool will log a message for every NSTask
that is spawned.  Included in each message is short description of the
task's purpose and a command-line expression that can be copied/pasted
into bash to replicate that command.

The output looks like --
https://gist.github.com/fpotter/6981876

I've wanted this for a long time - it's going to be super helpful for
debugging problems.