always pass the inverted test list to -SenTest / -XCTest
Instead of passing the explicit list of tests to be run to -SenTest, we
should instead pass a list of all the tests we do _NOT_ want to run and
set -SenTestInvertScope YES.

This is exactly how Xcode.app does it.  Also, for the Kiwi BDD testing
framework, this is the only way to run specific tests.  For some reason,
Kiwi will only run specific tests if we do the inversion method.