When analyzing, warn when build-state.dat is not in the expected location
Targets overriding the CONFIGURATION_TEMP_DIR build setting will emit
intermediate files in a non-standard location. Since xctool does not
parse the full build settings for sake of performance, it can't know
where the actual intermediates will be emitted to.

Previously, when collecting analyzer output, xctool blindly assumes the
location and existence of the file, and would block forever trying to
read it. Now it gives an insightful log message and ignores the target.