when we fail to start tests, include a placeholder test to advertise it.
We used to append "Test did not run: ..." message to every test in the
bundle that never got a chance to run because of some startup error.

Now we insert a fake placeholder test at the top.  This way, if the test
bundle is empty, we can still show a message about why it failed.  And,
that place holder tests can also include crash reports (if any).