don't use --adopt-pid when running the `sim` binary.
The adopt-pid option makes `sim` exec() the target executable instead of
just launching it and waiting for it to finish.

This introduces a very weird bug that's possible in the 10.8 kernel or
GCD where the launched process will hang indefinitely.  It cannot be
killed via -9, it cannot be attached to with the debugger.  It just sits
there, forever, doing nothing.  This only happens when running in
parallel mode, and even then, it doesn't happen every time.

From my investigation so far, the last code it ran before hanging was
something related to dispatch_semaphore stuff.

Removing --adopt-pid fixes the problem.