when running ios app tests, don't let stderr output get into the stream
This fixes the problem identified in:
https://github.com/facebook/xctool/issues/224#issuecomment-29288004

Normally, otest-shim would be capturing and rerouting STDERR to STDOUT
in packaged JSON objects.  But, in this case, the objc runtime was
emitting warnings to STDERR before otest-shim even loaded.

It's safe to just ignore all STDERR here when launching the sim.  The
only thing we'll be missing is whatever comes before otest-shim loads.