Set the same environment values for iOS App and Logic Test Query. Clean main.m of otest-query-ios.
We have found that sometimes we couldn't query tests because 'SenTest' or 'XCTest' frameworks start running them. When it happens we can see errors like this:

  Failed to query the list of test cases in the test bundle: Error while parsing JSON: The data has been corrupted.: {
      stderr = "2014-03-26 11:08:27.405 sim[8792:707] /Applications/xcode_5.0.2.app/Contents/Developer/Platforms/iPhoneSimulator.platform/usr/bin/sim: Simulator does not seem to be running, or may be running an old SDK.\nTest Suite 'All tests' started at 2014-03-26 18:08:28 +0000\nTest Suite 'All tests' finished at 2014-03-26 18:08:28 +0000.\nExecuted 0 tests, with 0 failures (0 unexpected) in 0.000 (0.002) seconds\n";
      stdout = "";
  }

Before we were trying to load the same code in otest-query-ios[osx] and otest-query-lib-ios[osx] but in some cases (especially when testing large code base) xctool wasn't able to query logic tests. It was found that by clearing main.m and not loading OtestQuery in it but dynamically loading otest-query-lib-ios[osx] as we are doing in case of application tests xctool could successfully query tests.