Improve simulated device and sdk selection mechanism.
Summary:
In xctool tests we are using cached output of xcodebuild -showbuildsettings. It could happen that sdk, that was originally used to generate it, i.e. iOS 6, is not available on the machine which actually runs xctool tests. In that case we could try to find the best compatible pair <simulated device, simulated iOS sdk>.