Always list tests and invert scope when running logic tests in Xcode 6.
For some reason if logic tests are run in parallel under Xcode 6 it happens that none of the tests are run in test bundle and only `{begin,end}-test-suite` events are populated to reporters.
Though the fix looks weird because now it will generate `-SenTest   -SenTestInvertScope YES` it seems to fix the issue and actually work.

The main suspicion is that it happens because of usage of `NSUserDefaults` in SenTesting/XCTest frameworks. Probably logic tests are overlapping at some point though it is not clear why it has started to happen only in Xcode 6.