Update otest-shim to work with Xcode 8
Since `dyld_register_image_state_change_handler` is not available
in Xcode 8, don't link against it and instead try to find it using
`dlsym`. If it is not available, instead swizzle
`-[NSBundle loadAndReturnError]` and swizzle the XCTest methods
after the bundle is loaded. The `XCTestLog` interface has changed
between Xcode 7 and 8, so this change also adds logic to detect
which interface is used and decide which methods to swizzle.

It is not my plan to get all of `xctool` working with Xcode 8, but
the shim might be useful on its own.