Close file descriptor only after IO channel clean up handler is called.
This addresses the following crash on Mojave:

```
BUG IN CLIENT OF LIBDISPATCH: Unexpected EV_VANISHED (do not destroy random mach ports or file descriptors)
```

dispatch_io_create "takes control of the specified file descriptor until one of the following occurs:
- You close the channel by calling the dispatch_io_close function.
- An unrecoverable error occurs on the file descriptor.
- All references to the channel are released.".

The system relinquishes control of the channel’s file descriptor only when the `cleanup_handler` is invoked.
So it was wrong to close IO channel file descriptor before the handler was invoked.